refactor(parser): unify BMS text decoding into a shared module
Extract decodeBmsText / decodeUtf8Text from parser.ts into
core/bms-text-decode.ts as the single bytes-to-text pipeline for every
runtime, and switch the web player's collection loader to it instead of
its local duplicate. Remove the unreachable iconv-lite based
bms-text-decoder.ts (never imported) together with the iconv-lite
dependency. Export decodeUtf8Text with a bench case and test.
